🔨 refactor: 블록 커스텀 코드 분리
